langgraph-demo/pyproject.toml
[project]
name = "financial-rag-agent"
version = "0.1.0"
description = "金融创业投资领域的智能问答系统，基于LangGraph的图RAG功能"
authors = [
    { name = "Financial RAG Team" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    # LangGraph 核心
    "langgraph>=1.0.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-sdk>=0.2.10",
    # 向量数据库支持（多数据库，高可用）
    "qdrant-client>=1.11.0",
    "pymilvus>=2.4.0",
    "weaviate-client>=4.0.0",
    # 文档解析
    "pypdf>=3.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "docx2python>=1.20",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "chardet>=5.0.0",
    # 文本处理
    "tiktoken>=0.7.0",
    # 前端
    "streamlit>=1.32.0",
    # 数据库和工具
    "psycopg[binary,pool]>=3.2.13",
    "python-dotenv>=1.0.1",
    # 其他工具
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest>=8.3.5"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"



[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.workspace]
members = [
    ".",
    ".",
    ".",
    ".",
    ".",
]

[tool.uv.sources]
financial-rag-agent = { workspace = true }

[tool.setuptools.package-dir]
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",
    "F",
    "I",
    "D",
    "D401",
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "financial-rag-agent",
]
